Shift threshold configurations into one function
