+++ src/utils/utils.js
@@ -1,7 +1,9 @@
-import { extent } from 'd3-array';
+import { max, extent } from 'd3-array';
 import { timeDay } from 'd3-time';
-import { timeFormat } from 'd3-time-format';
+import { timeFormat, utcParse } from 'd3-time-format';
+
 const formatDate = timeFormat('%Y-%m-%d');
+const parseDate = utcParse('%Y-%m-%d');
 
 export function buildScenarios(dataset) {
   // Instantiates constant scenarios used for a given geoid
@@ -19,6 +21,66 @@ export function buildScenarios(dataset) {
     scenarioArray.push(obj);
   }
   return scenarioArray;
+}
+
+export function configureThresholds(scenarioList, seriesList, allTimeDates, dateRange) {
+  // calculate smart default statThreshold, dateThreshold
+  // and return series with sims above thresholds and perc exceedences
+
+  const filteredSeriesList = []
+  const percExceedenceList = []
+
+  const idxMin = timeDay.count(allTimeDates[0], dateRange[0]);
+  const idxMax = timeDay.count(allTimeDates[0], dateRange[1]);
+
+  const filteredDates = Array.from(allTimeDates.slice(idxMin, idxMax));
+  const dateThresholdIdx = Math.ceil(filteredDates.length / 2)
+  const dateThreshold = filteredDates[dateThresholdIdx];
+  let statThreshold = 0;
+  let smartStatThresh = 0;
+  let sMin = Number.POSITIVE_INFINITY;
+  let sMax = 0;
+  let sliderMin = Number.POSITIVE_INFINITY;
+  let sliderMax = 0;
+
+  for (let i = 0; i < scenarioList.length; i++) {
+
+      [smartStatThresh, sliderMin, sliderMax] = calcSmartThresholds(
+          seriesList[i], i, statThreshold, idxMin, idxMax, sMin, sMax)
+
+      const simsOver = returnSimsOverThreshold(
+          seriesList[i], smartStatThresh, allTimeDates, dateThreshold);
+
+      // series has been mutated with above/below threshold are now added
+      const filteredSeries = filterByDate(seriesList[i], idxMin, idxMax)
+      filteredSeriesList.push(filteredSeries)
+
+      // calculate percExceedence based on series after filtering down
+      const percExceedence = filteredSeries.length > 0 ?
+          simsOver / filteredSeries.length : 0;
+      percExceedenceList.push(percExceedence)
+  }
+  return [filteredSeriesList, percExceedenceList, smartStatThresh, 
+          dateThreshold, sliderMin, sliderMax]
+}
+
+export function calcSmartThresholds(series, i, statThreshold, idxMin, idxMax, sliderMin, sliderMax) {
+  // calculates and returns default smart stat threshold
+  const filteredSeriesStat = filterByDate(series, idxMin, idxMax)
+
+  // array of all peaks in filtered series
+  const seriesPeaks = filteredSeriesStat.map(sim => max(sim.vals));
+  const [seriesMin, seriesMax] = getRange(seriesPeaks);
+
+  // ensures side by side y-scale reflect both series
+  if (seriesMin < sliderMin) sliderMin = seriesMin
+  if (seriesMax > sliderMax) sliderMax = seriesMax
+
+  // adds a range of "smart" value for statThreshold 
+  if (i === 0 && seriesMin < seriesMax / 2) statThreshold = seriesMin; 
+  if (i === 0 && seriesMin >= seriesMax / 2) statThreshold = seriesMax / 2; 
+
+  return [statThreshold, sliderMin, sliderMax];
 }
 
 export function returnSimsOverThreshold(series, statThreshold, dates, dateThreshold) {
@@ -42,6 +104,42 @@ export function returnSimsOverThreshold(series, statThreshold, dates, dateThresh
   return simsOver;
 }
 
+export function filterByDate(series, idxMin, idxMax) {
+  const filteredSeries = series.map( s => {
+      const newS = {...s}
+      newS.vals = s.vals.slice(idxMin, idxMax)
+      return newS
+  });
+  return filteredSeries
+}
+
+export function getConfBounds(dataset, scenarioList, severityList, stat, idxMin, idxMax) {
+  const confBoundsList = [];
+  for (let i = 0; i < scenarioList.length; i++) {
+      const confBounds = dataset[scenarioList[i].key][severityList[i].key][stat.key].conf;
+      const filteredConfBounds = confBounds.slice(idxMin, idxMax)
+      confBoundsList.push(filteredConfBounds);
+  }
+  return confBoundsList
+}
+
+export function getActuals(geoid, stat, scenarioList) {
+  const actualList = [];
+  // instantiate actuals data if data for specific indicator exists
+  for (let i = 0; i < scenarioList.length; i++) {
+    let actual = [];
+    const indicator = stat.name.toLowerCase();
+    const actualJSON = require('../store/actuals.json');
+    if (Object.keys(actualJSON).includes(indicator)) {
+        actual = actualJSON[indicator][geoid].map( d => {
+            return { date: parseDate(d.date), val: d.val}
+        });
+    }
+    actualList.push(actual);
+  }
+  return actualList
+}
+
 export function shuffle(array, numDisplaySims) {
   // returns randomly shuffled array of elements based on numDisplaySims
 
@@ -61,7 +159,6 @@ export function shuffle(array, numDisplaySims) {
     array[currIdx] = array[randomIdx]; 
     array[randomIdx] = tempVal; 
   }
-
   return array.slice(stopIdx, array.length);
 }
 
@@ -73,13 +170,13 @@ export function filterR0(series, r0selected, numDisplaySims) {
   const filtered = series.filter(s => s.r0 > r0min && s.r0 < r0max);
 
   // filter on numDisplaySims
-  const displaySims = shuffle(filtered.map(s => s.name), numDisplaySims);
+  const displaySims = shuffle(filtered.map(s => s.name), numDisplaySims); 
   const final = filtered.filter(s => displaySims.includes(s.name));
 
   return final;
 }
 
-export function createFilteredR0SeriesList(
+export function filterR0seriesList(
   r0selected, scenarioList, severityList, stat, dataset, numDisplaySims) {
   const r0FilteredSeriesList = []
   for (let i = 0; i < scenarioList.length; i++) {
